+++ core/components/romanesco/docs/changelog.md
@@ -7,6 +7,7 @@ New features:
 - Add post hook for pThumb to optimize the image and generate a WebP version
 
 Fixes and improvements:
+- Load all conditional assets through a single snippet
 - Add SUI list class to ul and ol tags in Markdown (only at first level)
 - Remove conditionals for empty logo paths (to clear media source path)
 - Add theme variables for inverted logo paths